+++ src/instrumentExplorer.ts
@@ -32,14 +32,6 @@ export class InstrumentsExplorer implements vscode.Disposable {
         Log.trace("Instantiating TDP", LOGLOC)
         const treeDataProvider = InstrumentProvider.instance
         Log.trace("Refreshing TDP", LOGLOC)
-        treeDataProvider
-            .refresh(async () => await this.startDiscovery())
-            .catch((e) => {
-                Log.error(
-                    `Problem starting Instrument List data provider: ${e}`,
-                    LOGLOC,
-                )
-            })
 
         this.InstrumentsDiscoveryViewer = vscode.window.createTreeView<
             Instrument | Connection | InactiveInstrumentList
@@ -49,18 +41,13 @@ export class InstrumentsExplorer implements vscode.Disposable {
 
         this.treeDataProvider = treeDataProvider
         vscode.commands.registerCommand("InstrumentsExplorer.refresh", () => {
-            this.InstrumentsDiscoveryViewer.message =
-                "Checking saved instrument connections..."
             this.treeDataProvider
                 ?.refresh(async () => {
                     await this.startDiscovery()
                 })
                 .then(
-                    () => {
-                        this.InstrumentsDiscoveryViewer.message = undefined
-                    },
+                    () => {},
                     (e) => {
-                        this.InstrumentsDiscoveryViewer.message = undefined
                         Log.error(`Unable to refresh instrument explorer: ${e}`)
                     },
                 )
@@ -91,17 +78,11 @@ export class InstrumentsExplorer implements vscode.Disposable {
         context.subscriptions.push(saveInstrument)
         context.subscriptions.push(removeInstrument)
 
-        this.InstrumentsDiscoveryViewer.message =
-            "Checking saved instrument connections..."
-
         this.treeDataProvider
             ?.refresh(async () => await this.startDiscovery())
             .then(
-                () => {
-                    this.InstrumentsDiscoveryViewer.message = undefined
-                },
+                () => {},
                 (e: Error) => {
-                    this.InstrumentsDiscoveryViewer.message = undefined
                     Log.error(`Unable to start Discovery ${e.message}`, LOGLOC)
                 },
             )
@@ -117,9 +98,6 @@ export class InstrumentsExplorer implements vscode.Disposable {
         }
         return new Promise<void>((resolve) => {
             if (!this._discoveryInProgress) {
-                this.InstrumentsDiscoveryViewer.message =
-                    "Discovering new instrument connections..."
-
                 this._discoveryInProgress = true
                 const discover = child.spawn(
                     DISCOVER_EXECUTABLE,
@@ -147,7 +125,6 @@ export class InstrumentsExplorer implements vscode.Disposable {
                     if (code) {
                         Log.trace(`Discover Exit Code: ${code}`, LOGLOC)
                     }
-                    this.InstrumentsDiscoveryViewer.message = ""
                     clearInterval(this.intervalID)
                     this._discoveryInProgress = false
                     resolve()