+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ### Fixed
 - Sometimes the incorrect instrument can appear to be renamed if a saved instrument is renamed
+- Stop instruments from shifting in Instrument Explorer view.
 - (**tsp-toolkit-kic-cli**) Reduce number of messages send to instrument when closing debugger
 - (**tsp-toolkit-kic-cli**) Terminal process panics when USB cable is removed during connection